refactoring: refactor and fix after merge
