--- src/index.html
@@ -1,134 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./reset.css">
-    <link rel="stylesheet" href="./style.css">
-    <title>RS-Leang Teams 165</title>
-    <link rel="icon" href="./assets/favicon.jpg" type="image/x-icon" />
-
-</head>
-
-<body class="body">
-    <div class="naw-wrapper">
-        <div class="burger-wrapper">
-            <div class="container-line-burger hide">
-                <div class="plase-burger">
-                    <svg class="svg-burger" focusable="false" viewBox="0 0 24 24" aria-hidden="true">
-                        <path d="M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z"></path>
-                    </svg>
-                </div>
-                <span class="style-h2 span-rs">RS Lang</span>
-            </div>
-            <ul class="ul-wrapper">
-                <li class="li-nav dark"><a href="#" class="a-nav">
-                        <div class="li-wrapper">
-                            <svg class="svg-nav" viewBox="0 0 24 24">
-                                <path d="M3 13h8V3H3v10zm0 8h8v-6H3v6zm10 0h8V11h-8v10zm0-18v6h8V3h-8z"></path>
-                            </svg>
-                            <span class="li-nav align-nav dark">Main</span>
-                        </div>
-
-                    </a></li>
-                <li class="li-nav"><a href="#" class="a-nav">
-                        <div class="li-wrapper">
-                            <svg class="svg-nav" focusable="false" viewBox="0 0 24 24" aria-hidden="true">
-                                <path
-                                    d="M12 11.55C9.64 9.35 6.48 8 3 8v11c3.48 0 6.64 1.35 9 3.55 2.36-2.19 5.52-3.55 9-3.55V8c-3.48 0-6.64 1.35-9 3.55zM12 8c1.66 0 3-1.34 3-3s-1.34-3-3-3-3 1.34-3 3 1.34 3 3 3z">
-                                </path>
-                            </svg>
-                            <span class="li-nav align-nav">Textbook</span>
-                        </div>
-                    </a></li>
-                <li class="li-nav hide"><a href="#" class="a-nav">
-                        <div class="li-wrapper">
-                            <svg class="svg-nav" focusable="false" viewBox="0 0 24 24" aria-hidden="true">
-                                <path
-                                    d="M18 2H6c-1.1 0-2 .9-2 2v16c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2V4c0-1.1-.9-2-2-2zM6 4h5v8l-2.5-1.5L6 12V4z">
-                                </path>
-                            </svg>
-                            <span class="li-nav align-nav">Dictionary</span>
-                        </div>
-                    </a></li>
-                <li class="li-nav"><a href="#" class="a-nav">
-                        <div class="li-wrapper">
-                            <svg class="svg-nav" focusable="false" viewBox="0 0 24 24" aria-hidden="true">
-                                <path
-                                    d="M21.58 16.09l-1.09-7.66C20.21 6.46 18.52 5 16.53 5H7.47C5.48 5 3.79 6.46 3.51 8.43l-1.09 7.66C2.2 17.63 3.39 19 4.94 19c.68 0 1.32-.27 1.8-.75L9 16h6l2.25 2.25c.48.48 1.13.75 1.8.75 1.56 0 2.75-1.37 2.53-2.91zM11 11H9v2H8v-2H6v-1h2V8h1v2h2v1zm4-1c-.55 0-1-.45-1-1s.45-1 1-1 1 .45 1 1-.45 1-1 1zm2 3c-.55 0-1-.45-1-1s.45-1 1-1 1 .45 1 1-.45 1-1 1z">
-                                </path>
-                            </svg>
-                            <span class="li-nav align-nav">Minigames</span>
-                        </div>
-                    </a></li>
-                <li class="li-nav"><a href="#" class="a-nav">
-                        <div class="li-wrapper">
-                            <svg class="svg-nav" focusable="false" viewBox="0 0 24 24" aria-hidden="true">
-                                <path d="M10 20h4V4h-4v16zm-6 0h4v-8H4v8zM16 9v11h4V9h-4z"></path>
-                            </svg>
-                            <span class="li-nav align-nav">Statistics</span>
-                        </div>
-                    </a></li>
-            </ul>
-        </div>
-    </div>
-    <div class="page-content">
-        <div class="header-wrapper">
-        </div>
-        </header>
-        <main class="main">
-            <div class="wrapper-man-and-lang">
-                <div class="svg-man-wrapper">
-                    <img class="svg-man img-saze-computer" src="./assets/man.svg" alt="svg">
-                </div>
-                <div class="rs-lang-wrapper">
-                    <span class="big-text">RS Lang</span>
-                    <span class="text blond" style="margin-top: 20px;">Learning English has never been so easy</span>
-                    <p class="text">Memorizing English words can be fun and challenging.<br> Play games, listen to
-                        pronunciation, improve your<br> knowledge. With our app, learning is a joy</p>
-                    <a href="https://rss-lang-team-53.web.app/textbook"><button class="button">let's start</button></a>
-                </div>
-            </div>
-            <div class="advantages-wrapper">
-                <h2 class="style-h2">Advantages</h2>
-                <div class="advantag-wrap"> </div>
-            </div>
-            <div class="posibilities-wrapper">
-                <h2 class="style-h2">All posibilities</h2>
-                <iframe class="video" src="https://www.youtube.com/embed/AD7eMN2phjc" title="YouTube video player"
-                    frameborder="0"
-                    allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                    allowfullscreen></iframe>
-            </div>
-
-            <div class="not-forget-wrapper">
-                <span class="big-text">Don't forget</span>
-                <div class="text-center">If you want to see statistics for all time and have <br> access to the
-                    dictionary, please login or sign up first.</div>
-                <div class="button-wrapper">
-                    <a href="https://rss-lang-team-53.web.app/textbook"><button
-                            class="button small new-color">login</button></a>
-                    <a href="https://rss-lang-team-53.web.app/textbook"><button class="button small">sing
-                            up</button></a>
-                </div>
-                <img class="img-saze-computer" src="./assets/computer.svg" alt="computer">
-            </div>
-            <div class="team-wrapper">
-                <h2 class="style-h2">Our team</h2>
-                <div class="our-team"></div>
-            </div>
-        </main>
-        <footer class="footer-wrapper">
-            <a href="https://rs.school/js/"><img src="./assets/rs_school_js.svg" alt="rs" width="90"></a>
-            <div class="name-link"></div>
-            <div class="footer-text">
-                <span class="text ">© 2022</span>
-            </div>
-        </footer>
-    </div>
-    <script type="module" src="./script.js"></script>
-</body>
-
-</html>
