+++ src/pages/components/arrAdvantages.ts
